* removing all 64 bit ints and functions * this version assumes you have a 16 bit cpu so that ints are 16 bits, and long ints are 32 bits * this version uses "%ld" to print 32 bit ints.

+++ json_object_private.h
@@ -30,7 +30,7 @@ struct json_object
   union data {
     boolean c_boolean;
     double c_double;
-    int64_t c_int64;
+    int32_t c_int;
     struct lh_table *c_object;
     struct array_list *c_array;
     struct {
